fix: added projectid state correction on / route for a new chat
